+++ CesiumVectorData/include/CesiumVectorData/GltfConverter.h
@@ -4,6 +4,7 @@
 #include <CesiumGeospatial/Ellipsoid.h>
 #include <CesiumGltf/Model.h>
 #include <CesiumGltf/Schema.h>
+#include <CesiumUtility/IntrusivePointer.h>
 #include <CesiumUtility/Result.h>
 #include <CesiumVectorData/GeoJsonDocument.h>
 
@@ -20,7 +21,8 @@ using ConverterResult = CesiumUtility::Result<CesiumGltf::Model>;
  * @brief The result of converting a MAXAR_content_geojson schema to a glTF
  * schema.
  */
-using ConvertSchemaResult = CesiumUtility::Result<CesiumGltf::Schema>;
+using ConvertSchemaResult =
+    CesiumUtility::Result<CesiumUtility::IntrusivePointer<CesiumGltf::Schema>>;
 
 /**
  * @brief Convert GeoJSON documents to glTF.
@@ -33,11 +35,16 @@ public:
    *
    * @param geoJson The GeoJSON document.
    * @param ellipsoid The ellipsoid for GeoJSON coordinates.
+   * @param pSchema the schema to use for converting GeoJSON feature properties
+   * to glTF EXT_structural_metadata properties.
    * @returns A result object that includes the glTF Model and any errors.
    */
   static ConverterResult convert(
       const GeoJsonDocument& geoJson,
-      const CesiumGeospatial::Ellipsoid& ellipsoid);
+      const CesiumGeospatial::Ellipsoid& ellipsoid,
+      const CesiumUtility::IntrusivePointer<CesiumGltf::Schema>& pSchema =
+          nullptr);
+
   /**
    * @brief Convert MAXAR_content_geojson schema into a glTF schema.
    * @param maxarSchema The schema extension object supplied with a GeoJson